+++ src/main/java/org/springframework/data/aerospike/core/ReactiveAerospikeTemplate.java
@@ -153,16 +153,22 @@ public class ReactiveAerospikeTemplate extends BaseAerospikeTemplate implements 
 
     @Override
     public <T> Flux<T> saveAll(Iterable<T> documents) {
-        validateForBatchWrite(documents, "Documents for saving");
-
+        String itemsDescription = "Documents for saving";
+        if (isInvalidForBatchWrite(documents, itemsDescription)) {
+            logEmptyItems(log, itemsDescription);
+            return Flux.empty();
+        }
         return saveAll(documents, getSetName(documents.iterator().next()));
     }
 
     @Override
     public <T> Flux<T> saveAll(Iterable<T> documents, String setName) {
         Assert.notNull(setName, "Set name must not be null!");
-        validateForBatchWrite(documents, "Documents for saving");
-
+        String itemsDescription = "Documents for saving";
+        if (isInvalidForBatchWrite(documents, itemsDescription)) {
+            logEmptyItems(log, itemsDescription);
+            return Flux.empty();
+        }
         return applyBufferedBatchWrite(documents, setName, SAVE_OPERATION);
     }
 
@@ -283,16 +289,22 @@ public class ReactiveAerospikeTemplate extends BaseAerospikeTemplate implements 
 
     @Override
     public <T> Flux<T> insertAll(Iterable<? extends T> documents) {
-        validateForBatchWrite(documents, "Documents for insert");
-
+        String itemsDescription = "Documents for inserting";
+        if (isInvalidForBatchWrite(documents, itemsDescription)) {
+            logEmptyItems(log, itemsDescription);
+            return Flux.empty();
+        }
         return insertAll(documents, getSetName(documents.iterator().next()));
     }
 
     @Override
     public <T> Flux<T> insertAll(Iterable<? extends T> documents, String setName) {
         Assert.notNull(setName, "Set name must not be null!");
-        validateForBatchWrite(documents, "Documents for insert");
-
+        String itemsDescription = "Documents for inserting";
+        if (isInvalidForBatchWrite(documents, itemsDescription)) {
+            logEmptyItems(log, itemsDescription);
+            return Flux.empty();
+        }
         return applyBufferedBatchWrite(documents, setName, INSERT_OPERATION);
     }
 
@@ -376,16 +388,22 @@ public class ReactiveAerospikeTemplate extends BaseAerospikeTemplate implements 
 
     @Override
     public <T> Flux<T> updateAll(Iterable<? extends T> documents) {
-        validateForBatchWrite(documents, "Documents for update");
-
+        String itemsDescription = "Documents for updating";
+        if (isInvalidForBatchWrite(documents, itemsDescription)) {
+            logEmptyItems(log, itemsDescription);
+            return Flux.empty();
+        }
         return updateAll(documents, getSetName(documents.iterator().next()));
     }
 
     @Override
     public <T> Flux<T> updateAll(Iterable<? extends T> documents, String setName) {
         Assert.notNull(setName, "Set name must not be null!");
-        validateForBatchWrite(documents, "Documents for update");
-
+        String itemsDescription = "Documents for updating";
+        if (isInvalidForBatchWrite(documents, itemsDescription)) {
+            logEmptyItems(log, itemsDescription);
+            return Flux.empty();
+        }
         return applyBufferedBatchWrite(documents, setName, UPDATE_OPERATION);
     }
 
@@ -482,50 +500,76 @@ public class ReactiveAerospikeTemplate extends BaseAerospikeTemplate implements 
 
     @Override
     public <T> Mono<Void> deleteAll(Iterable<T> documents) {
-        validateForBatchWrite(documents, "Documents for deleting");
-
+        String itemsDescription = "Documents for deleting";
+        if (isInvalidForBatchWrite(documents, itemsDescription)) {
+            logEmptyItems(log, itemsDescription);
+            return Mono.empty();
+        }
         return deleteAll(documents, getSetName(documents.iterator().next()));
     }
 
     @Override
     public <T> Mono<Void> deleteAll(Iterable<T> documents, String setName) {
         Assert.notNull(setName, "Set name must not be null!");
-        validateForBatchWrite(documents, "Documents for deleting");
-
+        String itemsDescription = "Documents for deleting";
+        if (isInvalidForBatchWrite(documents, itemsDescription)) {
+            logEmptyItems(log, itemsDescription);
+            return Mono.empty();
+        }
         return applyBufferedBatchWrite(documents, setName, DELETE_OPERATION).then();
     }
 
     @Override
     public <T> Mono<Void> deleteByIds(Iterable<?> ids, Class<T> entityClass) {
         Assert.notNull(entityClass, "Class must not be null!");
-        validateForBatchWrite(ids, "IDs");
+        String itemsDescription = "Ids for deleting";
+        if (isInvalidForBatchWrite(ids, itemsDescription)) {
+            logEmptyItems(log, itemsDescription);
+            return Mono.empty();
+        }
         return deleteByIds(ids, getSetName(entityClass));
     }
 
     @Override
     public <T> Mono<Void> deleteExistingByIds(Iterable<?> ids, Class<T> entityClass) {
         Assert.notNull(entityClass, "Class must not be null!");
-        validateForBatchWrite(ids, "IDs");
+        String itemsDescription = "Ids for deleting";
+        if (isInvalidForBatchWrite(ids, itemsDescription)) {
+            logEmptyItems(log, itemsDescription);
+            return Mono.empty();
+        }
         return deleteExistingByIds(ids, getSetName(entityClass));
     }
 
     @Override
     public Mono<Void> deleteByIds(Iterable<?> ids, String setName) {
         Assert.notNull(setName, "Set name must not be null!");
-        validateForBatchWrite(ids, "IDs");
+        String itemsDescription = "Ids for deleting";
+        if (isInvalidForBatchWrite(ids, itemsDescription)) {
+            logEmptyItems(log, itemsDescription);
+            return Mono.empty();
+        }
         return deleteByIds(ids, setName, false);
     }
 
     @Override
     public Mono<Void> deleteExistingByIds(Iterable<?> ids, String setName) {
         Assert.notNull(setName, "Set name must not be null!");
-        validateForBatchWrite(ids, "IDs");
+        String itemsDescription = "Ids for deleting";
+        if (isInvalidForBatchWrite(ids, itemsDescription)) {
+            logEmptyItems(log, itemsDescription);
+            return Mono.empty();
+        }
         return deleteByIds(ids, setName, true);
     }
 
     private Mono<Void> deleteByIds(Iterable<?> ids, String setName, boolean skipNonExisting) {
         Assert.notNull(setName, "Set name must not be null!");
-        validateForBatchWrite(ids, "IDs");
+        String itemsDescription = "Ids for deleting";
+        if (isInvalidForBatchWrite(ids, itemsDescription)) {
+            logEmptyItems(log, itemsDescription);
+            return Mono.empty();
+        }
 
         List<Object> idsList = new ArrayList<>();
         List<Mono<Void>> deleteResults = new ArrayList<>();
@@ -545,8 +589,11 @@ public class ReactiveAerospikeTemplate extends BaseAerospikeTemplate implements 
 
     private Mono<Void> doDeleteByIds(Collection<?> ids, String setName, boolean skipNonExisting) {
         Assert.notNull(setName, "Set name must not be null!");
-        validateForBatchWrite(ids, "IDs");
-
+        String itemsDescription = "Ids for deleting";
+        if (isInvalidForBatchWrite(ids, itemsDescription)) {
+            logEmptyItems(log, itemsDescription);
+            return Mono.empty();
+        }
         Key[] keys = ids.stream()
             .map(id -> getKey(id, setName))
             .toArray(Key[]::new);
@@ -555,8 +602,7 @@ public class ReactiveAerospikeTemplate extends BaseAerospikeTemplate implements 
     }
 
     private Mono<Void> batchDeleteAndCheckForErrors(IAerospikeReactorClient reactorClient, Key[] keys,
-                                                    boolean skipNonExisting)
-    {
+                                                    boolean skipNonExisting) {
         Function<BatchResults, Mono<Void>> checkForErrors = results -> {
             if (results.records == null) {
                 return Mono.error(new AerospikeException.BatchRecordArray(results.records,
@@ -579,11 +625,7 @@ public class ReactiveAerospikeTemplate extends BaseAerospikeTemplate implements 
 
     @Override
     public Mono<Void> deleteByIds(GroupedKeys groupedKeys) {
-        validateGroupedKeys(groupedKeys);
-
-        if (groupedKeys.getEntitiesKeys().isEmpty()) {
-            return Mono.empty();
-        }
+        if (areInvalidGroupedKeys(groupedKeys)) return Mono.empty();
 
         return deleteEntitiesByGroupedKeys(groupedKeys);
     }
@@ -888,11 +930,7 @@ public class ReactiveAerospikeTemplate extends BaseAerospikeTemplate implements 
 
     @Override
     public Mono<GroupedEntities> findByIds(GroupedKeys groupedKeys) {
-        validateGroupedKeys(groupedKeys);
-
-        if (groupedKeys.getEntitiesKeys().isEmpty()) {
-            return Mono.just(GroupedEntities.builder().build());
-        }
+        if (areInvalidGroupedKeys(groupedKeys)) return Mono.just(GroupedEntities.builder().build());
 
         return findGroupedEntitiesByGroupedKeys(reactorClient.getAerospikeClient()
             .copyBatchPolicyDefault(), groupedKeys);
